Flashlight update
